backend/db_fast_api/requirements.txt
aerich==0.4.2
aiosqlite==0.16.0
asyncpg==0.21.0
atlastk==0.12.2
beautifulsoup4==4.9.3
click==7.1.2
dnspython==2.0.0
email-validator==1.1.2
fastapi==0.62.0
google==3.0.0
h11==0.11.0
html5lib==1.1
idna==2.10
iso8601==0.1.13
pkg-resources==0.0.0
protobuf==3.14.0
pydantic==1.7.3
PyPika==0.44.1
pytz==2020.4
six==1.15.0
soupsieve==2.0.1
starlette==0.13.6
tortoise-orm==0.16.18
typing-extensions==3.7.4.3
uvicorn==0.12.3
webencodings==0.5.1
